Declaracao da Sequence na classe Pessoa. Ajuste no Repositorio. Criação da camada de serviço

+++ src/main/java/br/com/pamcary/repository/PessoaRepository.java
@@ -8,4 +8,4 @@ import br.com.pamcary.entity.Pessoa;
 
 
 @Repository
-public interface PessoaRepository extends JpaRepository<Pessoa, Long> { }
+public interface PessoaRepository extends JpaRepository<Pessoa, Integer> { }
